+++ crates/recent_projects/Cargo.toml
@@ -19,7 +19,6 @@ test-support = ["remote/test-support", "project/test-support", "workspace/test-s
 [dependencies]
 anyhow.workspace = true
 askpass.workspace = true
-auto_update.workspace = true
 db.workspace = true
 dev_container.workspace = true
 editor.workspace = true
@@ -29,7 +28,6 @@ fuzzy.workspace = true
 gpui.workspace = true
 language.workspace = true
 log.workspace = true
-markdown.workspace = true
 menu.workspace = true
 node_runtime.workspace = true
 open_path_prompt.workspace = true
@@ -39,6 +37,7 @@ picker.workspace = true
 project.workspace = true
 release_channel.workspace = true
 remote.workspace = true
+remote_connection.workspace = true
 semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -46,7 +45,6 @@ settings.workspace = true
 smol.workspace = true
 task.workspace = true
 telemetry.workspace = true
-theme.workspace = true
 ui.workspace = true
 util.workspace = true
 workspace.workspace = true
